+++ Introductory Problems/IncreasingArray.cpp
@@ -3,7 +3,7 @@ using namespace std;
 #define ll long long
 
 int main() {
-    ll n, p = INT32_MIN, element, ans = 0;
+    ll n, p = INT64_MIN, element, ans = 0;
     cin >> n;
     for (long i = 0; i < n; i++) {
         cin >> element;
